Re-include data extraction from COC and POP queries
In the process of clearing out code we removed the extraction of the relevant columns for the backing data. This means that the backing data now includes 3 variables we don't need to send: feedback_token, drug and drug_code.
This commit includes the extraction of relevant columns only.
To discuss with the Trusts if the need the columns that we've removed accidentally: "Taken.POP.before.","Ordered.POP.from.SH.24.before.","Ordered.OC.from.SH.24.before."