Adjusts COC and POP files for backing data
This commit re-adjusts the changes made in the COC and POP files we send to the Trusts to the previous specs. There are now more differences between the files we send to 4 different Trusts, however, we have been able to remove the 3 variables related to repeat users (OC taken before, ordered OC from SH24 before, and ordered COC/POP before).