Limits LLR STI backing data
This commit limits the data we send to LLR on STI activity to orders created from 1st January 2022. This is because the historical csv (from the start of the service) is now too heavy to send via nhs.net.
The change has been agreed with the Trust (Rob's email 7th Nov 2022).