Remove Site and Clinic from scripts
This commit removes the filters that use Site or Clinic from the scripts to extract data. These variables both related to the clinic, which have been removed form the database due to having no significance.
Most Trusts opposed to having changes to the backing data we submit monthly, as this impact their uploading systems. That's why we've included 'Clinic' and/or 'Site' as a blank or hard coded variable in the scripts.